Move breakpoint analysis entries to `breakpoints` reducer
