+++ src/devtools/client/debugger/src/reducers/breakpoints.ts
@@ -2,9 +2,11 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at <http://mozilla.org/MPL/2.0/>. */
 
-import { createSlice, PayloadAction } from "@reduxjs/toolkit";
-import { Location } from "@replayio/protocol";
+import { createSlice, createEntityAdapter, PayloadAction, EntityState } from "@reduxjs/toolkit";
+import { Location, PointDescription, AnalysisEntry } from "@replayio/protocol";
 import type { Context } from "devtools/client/debugger/src/reducers/pause";
+import string from "devtools/packages/devtools-reps/reps/string";
+import { AnalysisError } from "protocol/thread/analysis";
 import type { UIState } from "ui/state";
 
 import { getBreakpointsList } from "../selectors/breakpoints";
@@ -17,6 +19,37 @@ export type { Breakpoint } from "./types";
 
 type LocationWithoutColumn = Omit<Location, "column">;
 
+export enum AnalysisStatus {
+  // Happy path
+  Created = "Created",
+  LoadingPoints = "LoadingPoints",
+  PointsRetrieved = "PointsRetrieved",
+  LoadingResults = "LoadingResults",
+  Completed = "Completed",
+
+  // We don't have this yet, but we *should*, it's important. For instance, when
+  // a user changes their focus region and we're going to rerun this anyways?
+  // Cancel it!
+  Cancelled = "Cancelled",
+
+  // These errors mean very different things! The max hits for getting points is
+  // 10,000, while the max hits for running an analysis is 200!
+  ErroredGettingPoints = "ErroredGettingPoints",
+  ErroredRunning = "ErroredRunning",
+}
+
+export type AnalysisSummary = {
+  error: AnalysisError | undefined;
+  id: string;
+  location: Location | undefined;
+  condition?: string;
+  points: PointDescription[] | undefined;
+  // Do we want to store results here? I guess so? They can be a bit
+  // unweildy I think, but I expect they will be serializable.
+  results: AnalysisEntry[] | undefined;
+  status: AnalysisStatus;
+};
+
 export interface BreakpointsState {
   /**
    * Actual breakpoint entries, keyed by a location string
@@ -27,18 +60,25 @@ export interface BreakpointsState {
    * are still being processed by the Replay backend server
    */
   requestedBreakpoints: Record<string, LocationWithoutColumn>;
+  /**
+   * Analysis entries associated with breakpoints, keyed by GUID.
+   */
+  analyses: EntityState<AnalysisSummary>;
 }
+
+const analysesAdapter = createEntityAdapter<AnalysisSummary>();
 
 export function initialBreakpointsState(): BreakpointsState {
   return {
     breakpoints: {},
     requestedBreakpoints: {},
+    analyses: analysesAdapter.getInitialState(),
   };
 }
 
 const breakpointsSlice = createSlice({
   name: "breakpoints",
-  initialState: initialBreakpointsState(),
+  initialState: initialBreakpointsState,
   reducers: {
     setBreakpoint: {
       reducer(state, action: PayloadAction<{ breakpoint: Breakpoint; recordingId: string }>) {
@@ -89,10 +129,106 @@ const breakpointsSlice = createSlice({
       state.breakpoints = {};
       state.requestedBreakpoints = {};
     },
+
+    analysisCreated(state, action: PayloadAction<string>) {
+      analysesAdapter.addOne(state.analyses, {
+        error: undefined,
+        id: action.payload,
+        location: undefined,
+        condition: undefined,
+        points: undefined,
+        results: undefined,
+        status: AnalysisStatus.Created,
+      });
+    },
+    analysisPointsRequested(state, action: PayloadAction<string>) {
+      analysesAdapter.updateOne(state.analyses, {
+        id: action.payload,
+        changes: { status: AnalysisStatus.LoadingPoints },
+      });
+    },
+    analysisPointsReceived(
+      state,
+      action: PayloadAction<{ analysisId: string; points: PointDescription[] }>
+    ) {
+      analysesAdapter.updateOne(state.analyses, {
+        id: action.payload.analysisId,
+        changes: {
+          points: action.payload.points,
+          status: AnalysisStatus.PointsRetrieved,
+        },
+      });
+    },
+    analysisResultsRequested(state, action: PayloadAction<string>) {
+      analysesAdapter.updateOne(state.analyses, {
+        id: action.payload,
+        changes: { status: AnalysisStatus.LoadingResults },
+      });
+    },
+    analysisResultsReceived(
+      state,
+      action: PayloadAction<{ analysisId: string; results: AnalysisEntry[] }>
+    ) {
+      const { analysisId, results } = action.payload;
+
+      // Preemptively freeze the `results` array to keep Immer from recursively freezing,
+      // because we have mutable class instances nested inside (EW!)
+      Object.freeze(results);
+
+      analysesAdapter.updateOne(state.analyses, {
+        id: analysisId,
+        changes: { status: AnalysisStatus.Completed, results },
+      });
+    },
+    analysisErrored(
+      state,
+      action: PayloadAction<{
+        analysisId: string;
+        error: AnalysisError;
+        points?: PointDescription[];
+        results?: AnalysisEntry[];
+      }>
+    ) {
+      const { analysisId, error, points, results } = action.payload;
+
+      const analysis = state.analyses.entities[analysisId];
+      const currentStatus = analysis?.status;
+      const isLoadingPoints = currentStatus === AnalysisStatus.LoadingPoints;
+      const isLoadingResults = currentStatus === AnalysisStatus.LoadingResults;
+      if (!isLoadingPoints && !isLoadingResults) {
+        throw "Invalid state update";
+      }
+
+      if (points) {
+        Object.freeze(points);
+      }
+
+      if (results) {
+        Object.freeze(results);
+      }
+
+      analysesAdapter.updateOne(state.analyses, {
+        id: analysisId,
+        changes: {
+          status: isLoadingPoints
+            ? AnalysisStatus.ErroredGettingPoints
+            : AnalysisStatus.ErroredRunning,
+          points: isLoadingPoints ? points : analysis?.points,
+          results: isLoadingResults ? results : analysis?.results,
+          error,
+        },
+      });
+    },
   },
 });
 
 export const {
+  analysisCreated,
+  analysisErrored,
+  analysisPointsReceived,
+  analysisPointsRequested,
+  analysisResultsReceived,
+  analysisResultsRequested,
   removeBreakpoint,
   removeBreakpoints,
   removeRequestedBreakpoint,
